# raftpb/raft_service.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "RaftService",
    "version": "1.0",
    "contact": {
      "name": "RaftService",
      "url": "https://github.com/codeallergy/raft",
      "email": "zander@schwid.com"
    }
  },
  "tags": [
    {
      "name": "RaftService"
    }
  ],
  "schemes": [
    "http",
    "https"
  ],
  "consumes": [
    "application/json",
    "application/octet-stream"
  ],
  "produces": [
    "application/json",
    "application/octet-stream"
  ],
  "paths": {
    "/api/v1/fsm/recover": {
      "put": {
        "operationId": "RaftService_Recover",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "type": "object",
              "properties": {}
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/googlerpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "content",
            "in": "query",
            "required": false,
            "type": "string",
            "format": "byte"
          }
        ],
        "tags": [
          "RaftService"
        ]
      }
    },
    "/api/v1/raft/apply": {
      "get": {
        "operationId": "RaftService_ApplyCommand",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/raftStatus"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/googlerpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "payload",
            "in": "query",
            "required": false,
            "type": "string",
            "format": "byte"
          }
        ],
        "tags": [
          "RaftService"
        ]
      }
    },
    "/api/v1/raft/bootstrap": {
      "put": {
        "operationId": "RaftService_Bootstrap",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "type": "object",
              "properties": {}
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/googlerpcStatus"
            }
          }
        },
        "tags": [
          "RaftService"
        ]
      }
    },
    "/api/v1/raft/config": {
      "get": {
        "operationId": "RaftService_GetConfiguration",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/raftRaftConfiguration"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/googlerpcStatus"
            }
          }
        },
        "tags": [
          "RaftService"
        ]
      }
    },
    "/api/v1/raft/join": {
      "put": {
        "operationId": "RaftService_Join",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "type": "object",
              "properties": {}
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/googlerpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "nodeId",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "nodeAddr",
            "in": "query",
            "required": false,
            "type": "string"
          }
        ],
        "tags": [
          "RaftService"
        ]
      }
    }
  },
  "definitions": {
    "googlerpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    },
    "protobufAny": {
      "type": "object",
      "properties": {
        "@type": {
          "type": "string"
        }
      },
      "additionalProperties": {}
    },
    "raftRaftConfiguration": {
      "type": "object",
      "properties": {
        "state": {
          "type": "string"
        },
        "lastIndex": {
          "type": "string",
          "format": "uint64"
        },
        "serverList": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/raftRaftServer"
          }
        }
      }
    },
    "raftRaftServer": {
      "type": "object",
      "properties": {
        "nodeId": {
          "type": "string"
        },
        "raftAddr": {
          "type": "string"
        },
        "suffrage": {
          "type": "string"
        },
        "apiAddr": {
          "type": "string"
        }
      }
    },
    "raftStatus": {
      "type": "object",
      "properties": {
        "updated": {
          "type": "boolean"
        },
        "elapsed": {
          "type": "number",
          "format": "double",
          "title": "operation cost in seconds"
        },
        "id": {
          "type": "string",
          "title": "optional id field"
        }
      }
    }
  }
}